Improved deployment destruction
